+++ src/main/java/staysplit/hotel_reservation/review/domain/dto/request/CreateReviewRequest.java
@@ -1,10 +1,10 @@
 package staysplit.hotel_reservation.review.domain.dto.request;
 
 public record CreateReviewRequest(
-   Long id,
-   Long customerId,
+   Integer id,
+   Integer customerId,
    String nickname,
-   Long hotelId,
+   Integer hotelId,
    String content,
    Integer rating
 ) {}
