+++ src/main/java/staysplit/hotel_reservation/review/domain/dto/request/ModifyReviewRequest.java
@@ -1,7 +1,7 @@
 package staysplit.hotel_reservation.review.domain.dto.request;
 
 public record ModifyReviewRequest (
-        Long customerId,
+        Integer customerId,
         String content,
         Integer rating
 ){
